refactor(landing): reorganize components and extract constants
- Delete unused sparkles component and move form elements to constants
- Extract feature cards and values data to separate constants files
- Reorganize landing page components into proper folder structure
- Add "use client" directive to error pages
- Clean up footer imports and styling
- Improve code organization and maintainability by separating concerns
