+++ app/(main)/error.tsx
@@ -1,3 +1,4 @@
+"use client";
 import ErrorPage from '@/components/error/errorPage';
 import React from 'react'
