+++ app/(main)/unauthorized.tsx
@@ -1,3 +1,4 @@
+"use client";
 import UnauthorizedPage from '@/components/unauthorizedPage/unauthorizedPage';
 import React from 'react'
