+++ components/landing/about-us/constants.tsx
@@ -1,0 +1,22 @@
+import { Lightbulb, Target, Users } from "lucide-react";
+
+export const values = [
+        {
+            icon: Lightbulb,
+            title: "Innovation First",
+            description: "We push the boundaries of what's possible with AI-powered form technology.",
+            gradient: "from-cyan-500 to-blue-600",
+        },
+        {
+            icon: Users,
+            title: "User-Centric",
+            description: "Every feature we build starts with understanding our users' real needs.",
+            gradient: "from-blue-500 to-sky-600",
+        },
+        {
+            icon: Target,
+            title: "Results Driven",
+            description: "We measure success by the conversion rates and insights our users achieve.",
+            gradient: "from-sky-500 to-cyan-600",
+        },
+    ]
