+++ components/landing/hero/constants.tsx
@@ -1,0 +1,25 @@
+import { 
+    Type, 
+    CheckSquare, 
+    Calendar, 
+    Mail, 
+    Phone, 
+    BarChart3, 
+    PieChart, 
+    TrendingUp, 
+    Database, 
+    Brain, 
+} from "lucide-react";
+
+export const formElements = [
+    { icon: Type, label: "Text Input", color: "text-cyan-400" },
+    { icon: CheckSquare, label: "Checkbox", color: "text-sky-400" },
+    { icon: Calendar, label: "Date Picker", color: "text-cyan-500" },
+    { icon: Mail, label: "Email Field", color: "text-sky-500" },
+    { icon: Phone, label: "Phone Input", color: "text-cyan-400" },
+    { icon: BarChart3, label: "Analytics", color: "text-sky-400" },
+    { icon: PieChart, label: "Reports", color: "text-cyan-500" },
+    { icon: TrendingUp, label: "Insights", color: "text-sky-500" },
+    { icon: Database, label: "Data Flow", color: "text-cyan-400" },
+    { icon: Brain, label: "AI Processing", color: "text-sky-400" },
+]
