+++ components/landing/hero/floating-form-elements.tsx
@@ -1,19 +1,8 @@
 "use client"
 
-import { useEffect, useState } from "react"
-import { motion } from "framer-motion"
-import {
-  CheckSquare,
-  Type,
-  Calendar,
-  Mail,
-  Phone,
-  BarChart3,
-  PieChart,
-  TrendingUp,
-  Database,
-  Brain,
-} from "lucide-react"
+import { useEffect, useState } from "react";
+import { motion } from "framer-motion";
+import { formElements } from "./constants";
 
 export function FloatingFormElements() {
   const [dimensions, setDimensions] = useState({ width: 1200, height: 800 })
@@ -35,18 +24,7 @@ export function FloatingFormElements() {
     return () => window.removeEventListener("resize", handleResize)
   }, [])
 
-  const formElements = [
-    { icon: Type, label: "Text Input", color: "text-cyan-400" },
-    { icon: CheckSquare, label: "Checkbox", color: "text-sky-400" },
-    { icon: Calendar, label: "Date Picker", color: "text-cyan-500" },
-    { icon: Mail, label: "Email Field", color: "text-sky-500" },
-    { icon: Phone, label: "Phone Input", color: "text-cyan-400" },
-    { icon: BarChart3, label: "Analytics", color: "text-sky-400" },
-    { icon: PieChart, label: "Reports", color: "text-cyan-500" },
-    { icon: TrendingUp, label: "Insights", color: "text-sky-500" },
-    { icon: Database, label: "Data Flow", color: "text-cyan-400" },
-    { icon: Brain, label: "AI Processing", color: "text-sky-400" },
-  ]
+  
 
   return (
     <div className="absolute inset-0 overflow-hidden pointer-events-none" style={{ zIndex: 0 }}>
@@ -80,15 +58,15 @@ export function FloatingFormElements() {
             style={{ zIndex: 0 }}
           >
             <div className="relative group">
-              {/* Glow effect */}
+              
               <div className="absolute inset-0 bg-cyan-500/20 rounded-full blur-xl opacity-0 group-hover:opacity-100 transition-opacity duration-300"></div>
 
-              {/* Main element */}
+              
               <div className="relative w-14 h-14 bg-slate-900/40 backdrop-blur-sm rounded-lg border border-cyan-700/30 flex items-center justify-center shadow-lg">
                 <Icon className={`w-6 h-6 ${element.color}`} />
               </div>
 
-              {/* Tooltip */}
+              
               <div className="absolute -bottom-8 left-1/2 transform -translate-x-1/2 bg-slate-800/90 backdrop-blur-sm rounded px-2 py-1 text-xs text-slate-300 opacity-0 group-hover:opacity-100 transition-opacity duration-300 whitespace-nowrap border border-cyan-700/20">
                 {element.label}
               </div>
