+++ components/landing/landingPage.tsx
@@ -1,29 +1,28 @@
-"use client"
-
+"use client" 
 import Header from "@/components/header/header";
-import Hero from "@/components/landing/hero/hero";
-import FeaturesSection from "@/components/landing/features";
+import FeaturesSection from "@/components/landing/features/features";
 import CTASection from "@/components/landing/CTA";
-import AboutSection from "@/components/landing/about-us";
 import Footer from "@/components/landing/footer";
+import Hero from "./hero/hero";
+import AboutSection from "./about-us/about-us";
 
 export default function LandingPage() {
 
     return (
         <div className="flex min-h-screen flex-col mt-[-85px]">
-        <Header />
+            <Header/>
 
             <main className="flex-1">
                 <Hero/>
 
-                <FeaturesSection />
+                <FeaturesSection/>
 
-                <CTASection />
+                <CTASection/>
 
                 <AboutSection/>
             </main>
 
-        <Footer/>
+            <Footer/>
         </div>
     )
 }
